+++ phone-test/src/pages/CallDetail/CallDetail.decl.ts
@@ -1,0 +1,13 @@
+export interface CallDetailProps {
+    call: {
+      id: string;
+      direction: 'inbound' | 'outbound';
+      call_type: 'missed' | 'answered' | 'voicemail';
+      from: string;
+      to: string;
+      duration: number;
+      created_at: string;
+      notes?: string[];
+    };
+    onClick: (id: string) => void;
+  }
